+++ filter/RefUserFilter.js
@@ -18,7 +18,7 @@ module.exports = class RefUserFilter extends Base {
     async apply (query) {
         const user = query.controller.user.getId();
         const attr = query.view.class.getAttr(this.attr);
-        const ids = await attr.relation.refClass.find().and({[this.userAttr]: user}).ids();
+        const ids = await attr.relation.refClass.find({[this.userAttr]: user}).ids();
         return query.and({[this.attr]: ids});
     }
 };