+++ helper/MetaHelper.js
@@ -8,7 +8,7 @@ const Base = require('evado/component/helper/MetaHelper');
 module.exports = class MetaHelper extends Base {
 
     static resolveInteger (value, defaults, max = Number.MAX_SAFE_INTEGER) {
-        return !Number.isInteger(value) ? defaults : value > max ? max : value;
+        return !Number.isSafeInteger(value) ? defaults : value > max ? max : value;
     }
 
     static getRelatedMap (attr, models) {